# [A2-798] OPA update timing out -- part one: do less work