--- a/components/authz-service/engine/opa/opa.go
+++ b/components/authz-service/engine/opa/opa.go
@@ -321,9 +321,9 @@ func (s *State) V2IsAuthorized(
 	}
 }
 
-// V2ProjectsAuthorized evaluates whether a given [subject, resource, action] tuple
-// is authorized and returns the list of associated allowed projects
-// from the set of requested projects passed in.
+// V2ProjectsAuthorized evaluates whether a given [subject, resource, action,
+// projects] tuple is authorized and returns the list of associated allowed
+// projects from the set of requested projects passed in.
 func (s *State) V2ProjectsAuthorized(
 	ctx context.Context,
 	subjects engine.Subjects,