--- a/components/authz-service/server/v2/policy_refresher.go
+++ b/components/authz-service/server/v2/policy_refresher.go
@@ -188,13 +188,21 @@ func (refresher *policyRefresher) getPolicyMap(ctx context.Context) (map[string]
 
 		statements := make(map[string]interface{})
 		for _, st := range p.Statements {
-			statements[st.ID.String()] = map[string]interface{}{
-				"effect":    st.Effect.String(),
-				"role":      st.Role,
-				"projects":  st.Projects,
-				"actions":   st.Actions,
-				"resources": st.Resources,
+			stmt := map[string]interface{}{
+				"effect":   st.Effect.String(),
+				"projects": st.Projects,
 			}
+			// Only set these if provided
+			if st.Role != "" {
+				stmt["role"] = st.Role
+			}
+			if len(st.Actions) != 0 {
+				stmt["actions"] = st.Actions
+			}
+			if len(st.Resources) != 0 {
+				stmt["resources"] = st.Resources
+			}
+			statements[st.ID.String()] = stmt
 		}
 
 		members := make([]string, len(p.Members))