… components/authz-service/engine/opa/policy.bindata.go