@srenatus 2019-05-15
ℹ️ So, to be clear: this doesn't solve the issue; it's a byproduct of an ongoing investigation (hence the PR rename). Nonetheless, I'll merge this and keep moving this forward off of master in another PR.