@srenatus 2019-05-15
🤦‍♂ the benchmark I've added runs on an empty store 😄 willfix.